+++ lib/core_models/upload_file.model.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:file_picker/file_picker.dart';
 import 'package:image_picker/image_picker.dart';
 
-import 'package:genai_components/widgets/cl_media_viewer.widget.dart';
+import '../widgets/cl_media_viewer.widget.dart';
 
 class FFUploadedFile {
   final CLMedia clMedia;