+++ lib/pages/error_page.dart
@@ -1,10 +1,10 @@
-import 'package:genai_components/widgets/buttons/cl_button.widget.dart';
-import 'package:genai_components/providers/error_state.dart';
+import '../widgets/buttons/cl_button.widget.dart';
+import '../providers/error_state.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
-import 'package:genai_components/cl_theme.dart';
-import 'package:genai_components/layout/constants/sizes.constant.dart';
+import '../cl_theme.dart';
+import '../layout/constants/sizes.constant.dart';
 
 class ErrorPage extends StatefulWidget {
   final int? errorCode;