+++ lib/providers/app_state.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:rxdart/rxdart.dart';
-import 'package:genai_components/cl_theme.dart';
-import 'package:genai_components/utils/shared_manager.util.dart';
+import '../cl_theme.dart';
+import '../utils/shared_manager.util.dart';
 
 class AppState extends ChangeNotifier {
   late BehaviorSubject<bool> refreshList = BehaviorSubject<bool>.seeded(false);